requirements.txt
requests
lxml.html
selenium
